requirements.txt
aiogram
sqlalchemy
psycopg2-binary
python-decouple